Working prototype of a proxy
Yes, I know this is a massive commit, but you gotta understand this took me like a week to figure out what was going wrong. The problem was you can't use timeout on the read stream. It uses a read function that does not allow for canceling. The other fix was allowing for the ability to simultaneously read/write packets at the same time. This was achieved by using RwLock on structs that contained a MinecraftConnection, and then only use Mutexes on the TcpStreams themselves. Finally, the last fix was using 2 threads for the proxy forwarding stuff, since one thread could not keep up with all the packets coming upstream from the server.